Pokemon Codex v1.0
version 1.0 ready to go
