+++ web-page-internal/page-hierarchy/lessons/dokker-oppgaver/pokemon-codex/codex.html
@@ -10,20 +10,17 @@
   <link rel="icon" href="/content/icon/icon-original.svg">
 
   <!-- <script defer src="test-bench.js"></script> -->
-  <script defer src="codex-gamma.js"></script>
+  <script defer src="codex-gamma.js" type="module"></script>
 </head>
 <body>
 
-  <div id="target-cards"></div>
+  <div id="app-wrapper"></div>
 
-  <ul id="DEV-GUI">
+  <nav id="DEV-GUI">
     <input type="button" id="button-refresh" value="Refresh">
-    <input type="button" id="button-force-refresh" value="forceRefresh">
-    <!-- <input type="button" id="button-update-data" value="Update Data" disabled="true"> -->
-    <!-- <input type="button" id="button-update-details" value="Update Details" disabled="true"> -->
-    <input type="button" id="button-log-data" value="Log Data">
-    <!-- <input type="button" id="button-log-details" value="Log Detail" disabled="true"> -->
-  </ul>
+    <input type="button" id="button-create-gui" value="Create GUI">
+    <input type="button" id="button-delete-cards" value="Delete Cards">
+  </nav>
 
 </body>
 </html>
